Split state init
