+++ packages/client-p2p/src/peer/index.js
@@ -3,14 +3,11 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
+import type BN from 'bn.js';
 import type { LibP2P$Connection } from 'libp2p';
 import type PeerInfo from 'peer-info';
 import type { MessageInterface, PeerInterface } from '../types';
-import type { PeerState } from './types';
 
-const BN = require('bn.js');
-const EventEmitter = require('eventemitter3');
-const pushable = require('pull-pushable');
 const stringShorten = require('@polkadot/util/string/shorten');
 const l = require('@polkadot/util/logger')('p2p/peer');
 
@@ -18,18 +15,11 @@ const addConnection = require('./addConnection');
 const emitterOn = require('./emitterOn');
 const send = require('./send');
 const setBest = require('./setBest');
+const state = require('./state');
 
 module.exports = function createPeer (peerInfo: PeerInfo): PeerInterface {
   const id = peerInfo.id.toB58String();
-  const self: PeerState = {
-    bestHash: new Uint8Array([]),
-    bestNumber: new BN(0),
-    connections: [],
-    emitter: new EventEmitter(),
-    l,
-    nextId: 0,
-    pushable: pushable()
-  };
+  const self = state(l);
 
   return {
     id,
